Add inspector interface
This patch adds the interface that allows inspector clients to connect
and monitor all Segment activity happening on the page. This API is not
related to the events API (over Emitter interface) by design, as both of
these API's serve a very distinct purpose and will be developed at their
own pace without breaking each other's behaviour. End-users of the library
are not advised to use this interface in their production code and should
continue to rely on the existing events API.